package: PATH_PROVIDER static variable
